refactor: improve internal functions naming
